Increased Speed by 0.2
Increased the base and max speed of the player cells and cancer cells by 0.2. Makes it feel more snappy. Also tried to double the speed increases for the player and cancer cells, but that had a negative impact on the experience and made the cancer cells overwhelming, so that change was reverted.